+++ src/components/SessionEventContainer/SessionEventContainer.js
@@ -9,9 +9,12 @@ import { useStripes } from '../../StripesContext';
 import {
   RTR_ACTIVITY_CHANNEL,
   RTR_ERROR_EVENT,
+  RTR_FLS_TIMEOUT_EVENT,
+  RTR_FLS_WARNING_EVENT,
   RTR_TIMEOUT_EVENT
 } from '../Root/constants';
 import { toggleRtrModal } from '../../okapiActions';
+import FixedLengthSessionWarning from './FixedLengthSessionWarning';
 
 //
 // event listeners
@@ -25,9 +28,21 @@ export const thisWindowRtrError = (_e, stripes, history) => {
 };
 
 // idle session timeout in this window: logout
-export const thisWindowRtrTimeout = (_e, stripes, history) => {
+export const thisWindowRtrIstTimeout = (_e, stripes, history) => {
   stripes.logger.log('rtr', 'idle session timeout; logging out');
   history.push('/logout-timeout');
+};
+
+// fixed-length session warning in this window: logout
+export const thisWindowRtrFlsWarning = (_e, stripes, setIsFlsVisible) => {
+  stripes.logger.log('rtr', 'fixed-length session warning');
+  setIsFlsVisible(true);
+};
+
+// fixed-length session timeout in this window: logout
+export const thisWindowRtrFlsTimeout = (_e, stripes, history) => {
+  stripes.logger.log('rtr', 'fixed-length session timeout; logging out');
+  history.push('/logout');
 };
 
 // localstorage change in another window: logout?
@@ -113,6 +128,9 @@ const SessionEventContainer = ({ history }) => {
   // is the "keep working?" modal visible?
   const [isVisible, setIsVisible] = useState(false);
 
+  // is the fixed-length-session warning visible?
+  const [isFlsVisible, setIsFlsVisible] = useState(false);
+
   // inactivity timers
   const timers = useRef();
   const stripes = useStripes();
@@ -191,7 +209,7 @@ const SessionEventContainer = ({ history }) => {
       channels.window[RTR_ERROR_EVENT] = (e) => thisWindowRtrError(e, stripes, history);
 
       // idle session timeout in this window: logout
-      channels.window[RTR_TIMEOUT_EVENT] = (e) => thisWindowRtrTimeout(e, stripes, history);
+      channels.window[RTR_TIMEOUT_EVENT] = (e) => thisWindowRtrIstTimeout(e, stripes, history);
 
       // localstorage change in another window: logout?
       channels.window.storage = (e) => otherWindowStorage(e, stripes, history);
@@ -203,6 +221,13 @@ const SessionEventContainer = ({ history }) => {
       activityEvents.forEach(eventName => {
         channels.window[eventName] = (e) => thisWindowActivity(e, stripes, timers, bc);
       });
+
+      // fixed-length session: show session-is-ending warning
+      channels.window[RTR_FLS_WARNING_EVENT] = (e) => thisWindowRtrFlsWarning(e, stripes, setIsFlsVisible);
+
+      // fixed-length session: terminate session
+      channels.window[RTR_FLS_TIMEOUT_EVENT] = (e) => thisWindowRtrFlsTimeout(e, stripes, history);
+
 
       // add listeners
       Object.entries(channels).forEach(([k, channel]) => {
@@ -236,13 +261,19 @@ const SessionEventContainer = ({ history }) => {
     // array.
   }, []); // eslint-disable-line react-hooks/exhaustive-deps
 
-  // show the idle-session warning modal if necessary;
-  // otherwise return null
+  const renderList = [];
+
+  // show the idle-session warning modal?
   if (isVisible) {
-    return <KeepWorkingModal callback={keepWorkingCallback} />;
+    renderList.push(<KeepWorkingModal callback={keepWorkingCallback} />);
   }
 
-  return null;
+  // show the fixed-length session warning?
+  if (isFlsVisible) {
+    renderList.push(<FixedLengthSessionWarning />);
+  }
+
+  return renderList.length ? renderList : null;
 };
 
 SessionEventContainer.propTypes = {